Fix an error on name identification
When the first token could be a name or a surname, but with more chances to be a surname, I was inserting it into surnames list. This is wrong as I should presume that the first token will be a given name. I also added code to print correct confidence values